requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
asterisk-ami==0.1.7
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1.1
frozenlist==1.4.1
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-speech==2.27.0
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
multidict==6.1.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
pyst==0.9
requests==2.32.3
rsa==4.9
service-identity==24.1.0
setuptools==75.1.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.11.1
zope.interface==7.0.3
